# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CUENTY API - Enterprise
  description: |
    API REST completa para el sistema CUENTY de gestión de suscripciones para el mercado mexicano.
    
    ## Características
    - Autenticación JWT + API Keys
    - Rate limiting: 100 requests/minuto
    - Webhooks para integraciones
    - SDKs disponibles: Node.js, Python, PHP
    
    ## Ambiente
    - Base URL Producción: https://api.cuenty.com/v1
    - Base URL Sandbox: https://sandbox.api.cuenty.com/v1
    
  version: 3.0.0
  contact:
    name: CUENTY API Support
    email: api@cuenty.com
    url: https://cuenty.com/docs
  license:
    name: Propietario
    url: https://cuenty.com/legal/api-terms
servers:
  - url: https://api.cuenty.com/v1
    description: Servidor de Producción
  - url: https://sandbox.api.cuenty.com/v1
    description: Servidor de Sandbox
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para desarrolladores externos
  
  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    AnalyticsEvent:
      type: object
      required:
        - eventType
        - eventName
      properties:
        eventType:
          type: string
          example: page_view
        eventName:
          type: string
          example: home_viewed
        properties:
          type: object
          additionalProperties: true
        sessionId:
          type: string
          example: sess_123456
    
    # Webhook
...
paths:
...
  /analytics/events:
    post:
      tags:
        - Analytics
      summary: Registrar evento de analytics
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEvent'
      responses:
        '201':
          description: Evento registrado exitosamente
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Obtener métricas del dashboard
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Métricas del dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUsers:
                    type: integer
                  activeSubscriptions:
                    type: integer
                  totalRevenue:
                    type: number
                  conversionRate:
                    type: number
                  topStates:
                    type: array
                    items:
                      type: object
  /analytics/predictions/churn:
    get:
      tags:
        - Analytics
      summary: Obtener predicciones de churn
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Predicciones de churn
          content:
            application/json:
              schema:
                type: object
                properties:
                  predictions:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: integer
                        churnProbability:
                          type: number
                        confidenceScore:
                          type: number
  
  # ==========================================================================
  # WEBHOOKS
  # ==========================================================================
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []